RF+TST refactor tree_libs dict output
Change ``tree_libs`` output to use canonical (os.path.realpath) names
for both depending and depended libraries. The set of depending
libraries now becomes a dictionary with keys being the depending library
canonical names, and the values being the install name by which the
depending library refers to the depended library.

This means that we don't pick up the same library twice when there are
symlinks with different names pointing to the same library.