RF: root path does not have to exist
We were using /tmp as the root path that did not contain the libraries -
but that is confusing because, depending on what the /tmp directory
symlinks to, the test libraries might still be within /tmp.  Use a root
path that obviously doesn't exist to be more clear and less error-prone.