BF: drop irrelevant check for existing lib
I got confused.

When delocating, we check if a library is local to the code tree, or
not.  We need to modify the local libraries in-place to change their
install name paths.

We copy non-local libraries to `lib_path` before delocating.

For each local library, I was stupidly checking if there was a library
with the same name in `lib_path`, even though we are not going to copy
the local libararies, but modify them in-place.

Remove the check, and the tests for the check.