Replace invalidated signatures in copied libraries.
If a binary file is signed, tools like install_name_tool will invalidate
the signature by changing the file contents.  Later versions of MacOS
refuse to load binary files with invalid signatures.

This adds a new function which overwrites bad signatures with an ad-hoc
signature.  This is as close as you can get to removing the signature
using the available tools in MacOS.

This is tested on the rpath wheel, which was given a signature.