Process C extensions inside namespace packages
When a wheel contains subpackages of a namespace package, there is
no `__init__.py` file for the namespace package, though its
subpackage directories do contain `__init__.py` files.

When scanning the wheel for package directories, searching the
root for directories that contain `__init__.py` files is
insufficient because we will miss C extension modules that are
subpackages of namespace packages.

Instead, we need to do a depth-first traversal of the directory
tree and capture the most shallow directories that contain
`__init__.py` files.