Raise an exception on dependency resolution failures.
This only affects libraries with `@rpath` dependencies.  It's also me
fixing my own bad code when I wrote resolve_rpath.  At least I knew this
could be a problem back then and added a warning for it.

Previously this issue was warned about using Python's warning module,
but these errors are more often critical and will affect the
compatibility of the wheel.  A failure here means that either the
library actually is missing, delocate is not being correct, or the
library was built incorrectly.

The more I considered it the less I was able to justify the ability to
ignore these errors as an option since it's too implicit and could
easily ignore unintended dependencies.  I did add a parameter to
tree_libs which can be enabled to ignore these errors.