Show recursive dependencies in listdeps command.
Added tree_libs_from_directory function.  This is an improved version of
the tree_libs function.

The wheel_libs function was modified to use tree_libs_from_directory.
This is not fully backwards compatible since the new behavior
will now raise an error instead of returning an incomplete lib dict.