Remove TravisCI integration.
TravisCI used to be great, but now it is no longer capable of supporting
open-source projects like this one.
In its current state this CI only exists to cause a failure status.

Coverage is already handled by new workflows.
They also support `.coveragerc`.

The current workflows integrate with multibuild, and runs tests with the
same log levels.

Testing on older versions of MacOS is not be possible in general due to
a lack of support from Apple and GitHub for older MacOS versions.

Closes #145