Refactor mock tests with type hinting.
Conditional raises function changed to return context managers.
This is easier to work with than passing in a function each time.

The usage of nullcontext drops support for Python 3.6,
this version of Python has passed its end-of-life.

The collection of items used to make a mock test have been grouped into
a class to make them easier to work with.

This test has been parameterized instead of using a for loop.