Update retagging to use functions from delocating
This is now much simpler and more generalized. It first updates the name
to contain platform tags from both from_wheel and to_wheel. Next it
calls '_check_and_update_wheel_name' which fixes any issues with the
name AND will convert it to universal2 if necessary. And finally, it
calls '_update_wheelfile' to update the info in the WHEEL file to match
the new name.