Add another test for universal2 wheel version
Depending on the version of each wheel being fused, the universal2
wheel's version can be either the x86_64 or the arm64 wheel's version.
The first test for this is testing the scenario where the x86_64 wheel's
version should be used. This new test is tesing the scenario where the
arm64 wheel's version should be used.