Retag wheels automatically when fusing (#215)
* Add '--retag' flag to delocate-fuse command

This adds the ability to "retag" a universal2 fused wheel. When running
delocate-fuse with this flag, it will update the filename and the
dist-info file of the fused wheel to reflect that it is a universal2
wheel.

* Update _update_wheelfile to use pkginfo functions

* Detect if a wheel's name should use universal2

This updates '_get_archs_and_version_from_wheel_name' to check if both
arm64 and x86_64 platform tags are in the wheel's filename. If they are
both present, it changes the returned arch to be universal2 with the
appropriate version.

* Update retagging to use functions from delocating

This is now much simpler and more generalized. It first updates the name
to contain platform tags from both from_wheel and to_wheel. Next it
calls '_check_and_update_wheel_name' which fixes any issues with the
name AND will convert it to universal2 if necessary. And finally, it
calls '_update_wheelfile' to update the info in the WHEEL file to match
the new name.

* Simplify set comparison syntax

* Make fuse retagging the default behavior

This also updates the fuse function to support pathlib paths and updates
the retagging function to use pathlib paths.

* Update Changelog & README for retagging

* Update fuse_trees temp context

This also required some typing changes to functions called by
fused_trees.

* Add another test for universal2 wheel version

Depending on the version of each wheel being fused, the universal2
wheel's version can be either the x86_64 or the arm64 wheel's version.
The first test for this is testing the scenario where the x86_64 wheel's
version should be used. This new test is testing the scenario where the
arm64 wheel's version should be used.

* Move delocate-fuse functionality to delocate-merge

This moves the old 'delocate-fuse' functionality into the new
'delocate-merge' command. Running 'delocate-fuse' will now print a
message notifying the user of this change and then exit with an exit
code of 1. This also updates relevant tests, the changelog, and the
README.