Suppress macholib header parse error
Macholib raises ValueError when failing to parse files which can be used
to detect if the file is a Mach-O object.
It is also raised on Mach-O objects it doesn't know how to handle.

Checking for this error means `_is_macho_file` is redundant.

Add min version tests for various file types.
These tests clarify that static libraries don't have min version info.