Only add otool '-m' flag conditionally
Avoids unconditional usage of the `-m` flag which `llvm-otool` does not support.

Still leaves an edge case where `llvm-otool` is used AND the file contains
a parenthesis. A function testing if the current `otool` supports `-m`
would be more ideal but would be difficult for me to setup tests for.

Fixes #247